+++ OpenC-1.0/compiler/bootstrap/python/openc/__init__.py
@@ -1,3 +1,3 @@
 """OpenC 1.0 authored reference bootstrap compiler package."""
 
-__version__ = "1.0.0-rc.7"
+__version__ = "1.0.0-rc.8"